Move redux helpers into their own folder
