+++ src/currencyWallets/currency.test.js
@@ -6,7 +6,7 @@ import {
 } from '../test/fakeCurrency.js'
 import { fakeExchangePlugin } from '../test/fakeExchange.js'
 import { fakeUser, makeFakeAccount } from '../test/fakeUser.js'
-import { awaitState } from '../util/reaction.js'
+import { awaitState } from '../util/redux/reaction.js'
 import { assert } from 'chai'
 import { describe, it } from 'mocha'
 import { createStore } from 'redux'
