+++ src/redux/storageWallets/reducer.js
@@ -2,7 +2,7 @@ import {
   constReducer,
   listReducer,
   settableReducer
-} from '../../util/reducers.js'
+} from '../../util/redux/reducers.js'
 import { combineReducers } from 'redux'
 
 const ADD = 'airbitz-core-js/storageWallet/ADD'
