+++ src/util/redux/reactions.test.js
@@ -1,4 +1,4 @@
-import { makeAssertLog } from '../test/assertLog.js'
+import { makeAssertLog } from '../../test/assertLog.js'
 import { reactionMiddleware, createReaction, awaitState } from './reaction.js'
 import { assert } from 'chai'
 import { describe, it } from 'mocha'
